Rename package `enrichments` to singular
