+++ gradle-plugins/model-compiler/src/main/java/io/spine/gradle/compiler/ModelCompilerPlugin.java
@@ -22,7 +22,7 @@ package io.spine.gradle.compiler;
 import io.spine.gradle.compiler.annotation.ProtoAnnotatorPlugin;
 import io.spine.gradle.compiler.cleaning.CleaningPlugin;
 import io.spine.gradle.compiler.rejection.RejectionGenPlugin;
-import io.spine.gradle.compiler.lookup.enrichments.EnrichmentLookupPlugin;
+import io.spine.gradle.compiler.lookup.enrichment.EnrichmentLookupPlugin;
 import io.spine.gradle.compiler.lookup.proto.ProtoToJavaMapperPlugin;
 import io.spine.gradle.compiler.lookup.valrule.ValidationRulesLookupPlugin;
 import io.spine.gradle.compiler.validate.ValidatingBuilderGenPlugin;
